refactor(components): [carousel-item] (#10219)
