+++ packages/components/carousel/index.ts
@@ -13,4 +13,4 @@ export const ElCarouselItem = withNoopInstall(CarouselItem)
 export * from './src/carousel'
 export * from './src/carousel-item'
 
-export type { CarouselInstance } from './src/instance'
+export type { CarouselInstance, CarouselItemInstance } from './src/instance'
