+++ packages/components/carousel/src/carousel-item.ts
@@ -1,6 +1,5 @@
 import { buildProps } from '@element-plus/utils'
 import type { ExtractPropTypes } from 'vue'
-import type CarouselItem from './carousel-item.vue'
 
 export const carouselItemProps = buildProps({
   name: { type: String, default: '' },
@@ -11,5 +10,3 @@ export const carouselItemProps = buildProps({
 } as const)
 
 export type CarouselItemProps = ExtractPropTypes<typeof carouselItemProps>
-
-export type CarouselItemInstance = InstanceType<typeof CarouselItem>
