+++ packages/components/carousel/src/instance.ts
@@ -1,3 +1,5 @@
 import type Carousel from './carousel.vue'
+import type CarouselItem from './carousel-item.vue'
 
 export type CarouselInstance = InstanceType<typeof Carousel>
+export type CarouselItemInstance = InstanceType<typeof CarouselItem>
